requirements.txt
# 非均质有雾图像去雾应用依赖包
# GUI框架
PyQt6>=6.5.0

# 图像处理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.10.0
scikit-image>=0.20.0

# 数据处理和工具
dataclasses-json>=0.5.9
psutil>=5.9.0

# 打包工具
PyInstaller>=5.10.0

# 开发工具（可选）
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0